Bytt ut unødvendig requestlogger med logger med timing. Forenkle config noe.
